Fix volume slides with 00 parameter
from:	 Mirko Buffoni <mirbuf@gmail.com>
to:	 Claudio Matsuoka <cmatsuoka@gmail.com>
date:	 Mon, Apr 27, 2009 at 7:29 AM
subject: A00 effect problem

Ref. Guild of Sounds.xm

here you can hear the problem.  At the very end of the module, there
is a bass sound which should be fading out, while in XMP is not.

The sound is played and set the effect A01, and later a sequence of
A00 should repeat the previous effect, fading out the sound.  Since
there are holes between A01 and A00, maybe XMP is forgetting the
previous value of the effect.

I have other modules, but most of the problems seems related to the
fact the previous effect is not played when value = 00.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>