Add unified pitch slide/portamento quirk
http://eval.sovietrussia.org/wiki/Player_abuse_tests

Compatible Gxx off

Impulse Tracker links the effect memories for Exx, Fxx, and Gxx together if "Compatible Gxx" in NOT enabled in the file header. In other formats, portamento to note is entirely separate from pitch slide up/down. Several players that claim to be IT-compatible do not check this flag, and always store the last Gxx value separately.

When this test is played correctly, the note will bend up, down, and back up again. Players which do not correctly handle the Compatible Gxx flag will not perform the final pitch slide.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>