Add generalized delay with IT instrument quirk
From Storlek test #8:
http://eval.sovietrussia.org/wiki/Player_abuse_tests#Out-of-range_note_delays

This test is to make sure note delay is handled correctly if the delay
value is out of range. The correct behavior is to act as if the entire
row is empty. Some players ignore the delay value and play the note
on the first tick.

Oddly, Impulse Tracker does save the instrument number, even if the
delay value is out of range. I'm assuming this is a bug; nevertheless,
if a player is going to claim 100% IT compatibility, it needs to copy
the bugs as well.

When played correctly, this should play the first three notes using
the square wave sample, with equal time between the start and end of
each note, and the last note should be played with the noise sample.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>