Digital Symphony fixes by Tom Hargreaves
from	Tom Hargreaves <hex@freezone.co.uk>

While the only difference between v0 and v1 is apparently that
the latter supports extra compression formats, I've seen v1 files
that don't use any formats not in v0, and that play just fine if
the version check is removed.  Is returning "unrecognised file
format" mid-way through the samples after a successful initial
identification acceptable for other v1 files?  It seems silly not
to try to play them.  Also the existing loader for compression
format 4 is terribly wrong and loses sync, so I've removed it.

Happily I have two original playroutines, one that reads only v0
and one that can read v1 too, and have started a spot of
comparative disassembly in the hope of finding how to decompress
v1.

The finetune shift is very definitely 4 and not 3, both from
the docs I have and going by the tune "Project X" (title
"px.bladswede remix!") which you've already tried judging from the
CVS log.

Test tunes: Boombass is a v1 file with no sample compression.
(While these samples look too small to compress, I have another v1
file with no compression for which this is not the case.)

Klisje, one of the demo files provided with Digital Symphony
itself, has disabled effects.  I have a theory that the
whole "allowed effects table" bodge was a result of this mod's
use of speed 0x20.  While this file doesn't play identically to
how the original soundtracker is intended to, the patch makes it
play like it does in Digital Symphony.

Drumbass tests tempo change commands.  This file has no sample
names or compression because it was ripped from a
tune-plus-player-in-a-single-executable (whatever the word for
one of those is).

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>