Make zipfile detection stricter
The zip file format starts with 0x50 0x4B 0x03 0x04, according to the
pkware specifications.  This stricter test is necessary because
old-school noisetracker mods inclde the filename in the file, which can
easily start with 'PK', which is the case for the Star Control 2 song
'PKUNK.MOD'.  xmp tries to treat it as a zipfile, which obviously fails.
(from Solomon Peachy)

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>