Don't block if running in background
(from Ralf Hoffmann)
I noticed a problem with the background detection and keyboard handling.
If xmp is started as a background job it won't start playing since the
code wants to read input from the keyboard which will however block
forever. It will block the same way even if it has been started in the
foreground by another script (see below). I have attached a small patch
which catches this particular problem.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>