Remove nonexistent Modplug Tracker IT quirk
from Johannes Schultz <info@sagamusix.de>
subject	XMP quirks and OpenMPT identification

I've noticed a strange "fix" in XMP which makes zero sense to me:
"added ModPlug Tracker IT quirk: ignore sample global volume (fixes
speech in "Deep In Her Eyes Remake", reported by Douglas Carmichael)"
This is wrong. The speech sample does not play in ModPlug either
(not in MPT 1.09, not in MPT 1.16 and also not in OpenMPT). The bug
report you got there was "wrong". The global volume of the speech sample
is supposed to be 0, not 64.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>