Add Audacious 3.0 plugin
Audacious 3.0 "[...] will be removing the ::configure() and ::about()
functions for plugins and forcing use of the PluginPreferences [...]":

 http://boards.audacious-media-player.org/viewtopic.php?f=1&t=364

At Fedora I've upgrade Audacious to 3.0-beta1 meanwhile, and in order
to be able to build the XMP plugin again, I've modified my previous
port to Audacious 2.5 even further to use the PreferencesWidget API.

With regard to playback, no new changes have been necessary.

TODO: Display the list of supported formats. This used to be available
in the "About" dialog, and I haven't worked on that due to lack of time
and because I may need to review the About dialog GTK usage eventually
anyway.

Signed-off-by: Michael Schwendt <mschwendt@gmail.com>
Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>