Fix AMF loader to include empty track 0
Also assign track 0 to any invalid track in the index.

Andre Timmermans description:

Order table: the track numbers in this table are not explained, but as
you noticed  you have to perform -1 to obtain the index in the track table.
For value 0, found in some files, I think it means an empty track.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>