Disable "XM old instrument" quirk
Benjamin Shadwick <benshadwick@gmail.com> alerts me that vestiges.xm
is playing incorrectly at 0:51. Cause is a note that should be cut
by a new note with invalid instrument but isn't because the "old instrument"
quirk is active. Tested against FT2 and verified that this quirk seems
to be incorrect, so disabled it here.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>