xmp: update tty state as soon as we are in the foreground
When xmp is _running_ in the background, and then is brought to the
foreground, the program will not necessarily be notified of this state
change from external sources, as shells generally do not seem to send
gratitious SIGCONTs.

Therefore, to get keyboard input and status updates working again once
xmp is in the foreground, xmp actively needs to monitor the pgrps
itself, which this patch implements. (rsync does it the same way.)